requirements.txt
# Core dependencies
streamlit>=1.28.0
python-dotenv>=1.0.0

# AI and ML
pydantic>=2.0.0
pydantic-ai==1.0.0
pydantic-graph

# OpenAI and Google
openai>=1.0.0
google-genai>=0.1.0

# Web scraping and search
tavily-python>=0.3.0
langchain-community>=0.0.1
beautifulsoup4>=4.12.0
requests>=2.31.0

# Logging and monitoring
logfire>=0.1.0

# Presentation generation
python-pptx>=0.6.21
Pillow>=10.0.0

# Data processing
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

